refactored pathfinding function into seperate file
