+++ server/config.js
@@ -1,4 +1,4 @@
-let config = {
+const config = {
     host: 'sql.tabletopartistry.net',
     port: '8100',
     user: 'tunnelapp-user',
